BismillahFiks/publisher/requirements.txt
paho-mqtt>=1.6.0
requests
python-dotenv
schedule